Move lightbox classes to Lightboxes namespace
