+++ R7.MiniGallery/R7.MiniGallery.csproj
@@ -95,6 +95,7 @@
     <Folder Include="Views\Shared\" />
     <Folder Include="Controllers\" />
     <Folder Include="ViewModels\" />
+    <Folder Include="Lightboxes\" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="module.css" />
@@ -172,10 +173,6 @@
       <DependentUpon>BulkAdd.ascx</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="components\Colorbox.cs" />
-    <Compile Include="components\Lightbox.cs" />
-    <Compile Include="components\LightboxBase.cs" />
-    <Compile Include="components\NoneBox.cs" />
     <Compile Include="components\Utils.cs" />
     <Compile Include="components\MiniGallerySettingsRepository.cs" />
     <Compile Include="Data\MiniGalleryDataProvider.cs" />
@@ -184,6 +181,10 @@
     <Compile Include="Controllers\ImageController.cs" />
     <Compile Include="Models\IImage.cs" />
     <Compile Include="ViewModels\ImageViewModel.cs" />
+    <Compile Include="Lightboxes\Colorbox.cs" />
+    <Compile Include="Lightboxes\Lightbox.cs" />
+    <Compile Include="Lightboxes\LightboxBase.cs" />
+    <Compile Include="Lightboxes\NoneBox.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="$(MSBuildExtensionsPath)\Microsoft\VisualStudio\v10.0\WebApplications\Microsoft.WebApplication.targets" />
